+++ Stocks/Modules/News/NewsVM.swift
@@ -43,7 +43,7 @@ class NewsVM: NewsViewModel {
         }
     }
     
-    func fetchImage(from url: URL, withSize size: Float, for indexPath: IndexPath) {
+    func fetchImage(from url: URL, withSize size: Double, for indexPath: IndexPath) {
         cacheService.fetchImage(from: url, withSize: size) { [weak self] image in
             DispatchQueue.main.async {
                 self?.view?.showImage(image, at: indexPath)
