+++ Stocks/Modules/Stocks/StocksContract.swift
@@ -37,7 +37,7 @@ protocol StocksViewModel: AnyObject {
     var watchlist: [CompanyProfile] { get }
     
     func updateWatchlist(at index: Int, to newIndex: Int?, with action: Action)
-    func fetchLogo(from url: URL, withSize size: Float, for indexPath: IndexPath)
+    func fetchLogo(from url: URL, withSize size: Double, for indexPath: IndexPath)
     func fetchQuotes(for company: CompanyProfile, at indexPath: IndexPath)
     func onlineUpdateBegin()
     func onlineUpdateEnd()
